Fix old bug reporting wrong capacity for SD Cards
The version has been bumped up to 4.8.5.

The max number of 512B sectors was never reported correctly.
It has been causing the board not working when some SCSI
harddrives are set to the max reported capacity of the SD Card.
As a result of the fix, the max capacity is higher than
the erronous max capacity with some SD cards.

Besides the fix, the SCSI util code has been removed from the repo.