Fix SD Card block capacity reading
Firmware bumped to version 5.2.1

The max number of blocks for an SD card was read incorrectly. When a
SCSI device is set to the max number of blocks some SD cards can contain
would lock up the board.  Fixing the issue gives a larger usable max
capacity with some cards and the board no longer locks when set to
the limit.

Also deleted PSoC Creator projects for everyting but the V5.2 boards
and the scsi2sd util code.